+++ server/src/js/components/map/button.js
@@ -1,0 +1,21 @@
+
+export function createMapButton (buttonContainer, uri, options) {
+  const mapId = buttonContainer.id
+
+  // Create map button
+  const button = document.createElement('a')
+  button.setAttribute('href', uri)
+  if (options.btnType !== 'link') {
+    button.setAttribute('role', 'button')
+    button.setAttribute('data-module', 'govuk-button')
+  }
+  button.id = mapId + '-btn'
+  button.innerHTML = `<span class="defra-button-secondary__icon"><svg focusable="false" width="15" height="20" viewBox="0 0 15 20"><path d="M15,7.5c0.009,3.778 -4.229,9.665 -7.5,12.5c-3.271,-2.835 -7.509,-8.722 -7.5,-12.5c0,-4.142 3.358,-7.5 7.5,-7.5c4.142,0 7.5,3.358 7.5,7.5Zm-7.5,5.461c3.016,0 5.461,-2.445 5.461,-5.461c0,-3.016 -2.445,-5.461 -5.461,-5.461c-3.016,0 -5.461,2.445 -5.461,5.461c0,3.016 2.445,5.461 5.461,5.461Z" fill="currentColor"/></svg></span><span class="defra-button-secondary__text">${options.btnText || 'View map'}</span><span class="govuk-visually-hidden">(Visual only)</span>`
+  button.className = options.btnClass || (options.btnType === 'link' ? 'defra-link-icon-s' : 'defra-button-secondary')
+
+  if (options.data && options.data.button) {
+    button.setAttribute('data-journey-click', options.data.button)
+  }
+
+  return button
+}
