+++ server/src/js/components/map/live.js
@@ -13,6 +13,7 @@ import { Point, MultiPolygon } from 'ol/geom'
 import { buffer, containsExtent, getCenter } from 'ol/extent'
 import { Vector as VectorSource } from 'ol/source'
 import moment from 'moment-timezone'
+import { createMapButton } from './button'
 
 const { addOrUpdateParameter, getParameterByName, forEach } = window.flood.utils
 const maps = window.flood.maps
@@ -690,18 +691,7 @@ maps.createLiveMap = (mapId, options = {}) => {
 
   // Create map button
   const btnContainer = document.getElementById(mapId)
-  const button = document.createElement('a')
-  button.setAttribute('href', uri)
-  if (options.btnType !== 'link') {
-    button.setAttribute('role', 'button')
-    button.setAttribute('data-module', 'govuk-button')
-  }
-  button.id = mapId + '-btn'
-  button.innerHTML = `<span class="defra-button-secondary__icon"><svg focusable="false" width="15" height="20" viewBox="0 0 15 20"><path d="M15,7.5c0.009,3.778 -4.229,9.665 -7.5,12.5c-3.271,-2.835 -7.509,-8.722 -7.5,-12.5c0,-4.142 3.358,-7.5 7.5,-7.5c4.142,0 7.5,3.358 7.5,7.5Zm-7.5,5.461c3.016,0 5.461,-2.445 5.461,-5.461c0,-3.016 -2.445,-5.461 -5.461,-5.461c-3.016,0 -5.461,2.445 -5.461,5.461c0,3.016 2.445,5.461 5.461,5.461Z" fill="currentColor"/></svg></span><span class="defra-button-secondary__text">${options.btnText || 'View map'}</span><span class="govuk-visually-hidden">(Visual only)</span>`
-  button.className = options.btnClasses || (options.btnType === 'link' ? 'defra-link-icon-s' : 'defra-button-secondary defra-button-secondary--icon')
-  if (options.data && options.data.button) {
-    button.setAttribute('data-journey-click', options.data.button)
-  }
+  const button = createMapButton(btnContainer, uri, options)
   btnContainer.parentNode.replaceChild(button, btnContainer)
 
   // Detect keyboard interaction
